+++ Forge/changelog.md
@@ -6,3 +6,4 @@ TODO:
 - mod.azure.azurelib.render.layer.AzArmorLayer
 - mod.azure.azurelib.render.textures.AnimatableTexture
 - mod.azure.azurelib.render.textures.AzAbstractTexture
+- mod.azure.azurelib.render.textures.AzAutoGlowingTexture